Refactor item pickup and drop functionality
