+++ Assets/Art/Sprites.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e59befdce10e9d748bc83c97ec17e95b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
