+++ Assets/Prefabs/ToolRollSlot.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b8ea0a92cede88043a0fe8ff74ebe051
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
