+++ Assets/Scripts/ToolController.cs
@@ -1,0 +1,72 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.EventSystems;
+using Tool;
+
+public class ToolController : MonoBehaviour, IPointerDownHandler, ITool {
+    [SerializeField] private Canvas canvas;
+    [SerializeField] private ToolType toolType; 
+
+    private RectTransform rectTransform;
+    private CanvasGroup canvasGroup;
+    private Vector3 mousePosition;
+    private Collider2D targetCollider;
+
+    private bool pickedUp;
+
+    private void Awake() {
+        pickedUp = false;
+        rectTransform = GetComponent<RectTransform>();
+        canvasGroup = GetComponent<CanvasGroup>();
+    }
+
+    private void Update() {
+        if (Camera.main.ScreenToWorldPoint(Input.mousePosition) != mousePosition) {
+            mousePosition = Camera.main.ScreenToWorldPoint(Input.mousePosition);
+        }
+
+        if (pickedUp) {
+            rectTransform.position = Input.mousePosition;
+        }
+    }
+
+    public void OnPointerDown(PointerEventData eventData) {
+        // Check if the tool needs to be picked up
+        if (!pickedUp) {
+            pickedUp = true;
+            canvasGroup.alpha = .6f;
+            canvasGroup.blocksRaycasts = false;
+        } else {
+            // Check if we are above a damage spot
+            targetCollider = Physics2D.OverlapPoint(mousePosition);
+
+            // Check if over plushy
+            if (targetCollider.transform.gameObject.tag == "Damage") {
+                ApplyTool();
+            }
+        }
+    }
+
+    public void ApplyTool() {
+        // On click, check if the collider is a valid damage type for the selected tool
+    }
+
+    // Drop tool back on starting position
+    private void DropTool() {
+        pickedUp = false;
+    }
+
+    public bool PickedUp {
+        get { return pickedUp; }
+    }
+
+    private void OnEnable() {
+        EventManager.StartListening("DropTool", DropTool);
+    }
+
+    private void OnDisable() {
+        EventManager.StopListening("DropTool", DropTool);
+    }
+
+}
